Distributed tests T0003
